Allow using custom bulk actions component for replay search.
