+++ graylog2-web-interface/src/components/events/bulk-replay/types.ts
@@ -1,0 +1,6 @@
+import type { Event } from 'components/events/events/types';
+
+export type RemainingBulkActionsProps = {
+  events: Event[];
+  completed: boolean;
+}
